pyproject.toml
[tool.poetry]
name = "certificates"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pypdf2 = "^3.0.1"
pymupdf = "^1.23.22"
reportlab = "^4.1.0"
pillow = "^10.2.0"
openpyxl = "^3.1.2"
progress = "^1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
